server/backend/requirements.txt
flask
flask-cors
torch
numpy
pandas
google-genai
python-dotenv
groq
psycopg2-binary
